+++ src/player-util.h
@@ -103,6 +103,8 @@ bool player_can_study_prereq(void);
 bool player_can_read_prereq(void);
 bool player_can_fire_prereq(void);
 bool player_can_refuel_prereq(void);
+bool player_used_cheat_option(const struct player *p);
+bool player_has_cheated(const struct player *p);
 bool player_can_debug_prereq(void);
 bool player_book_has_unlearned_spells(struct player *p);
 bool player_confuse_dir(struct player *p, int *dir, bool too);